refactor: ExpenseFixture 중복 파일 통합
